refactor: lib queries and remove unused utilities
